Reimplements Guided Conversation lib as a skill.
