--- libraries/python/skills/skills/form-filler-skill/form_filler_skill/definition.py
@@ -1,14 +1,0 @@
-from dataclasses import dataclass
-from typing import Optional
-
-from form_filler_skill.resources import ResourceConstraint
-
-
-@dataclass
-class GCDefinition:
-    artifact_schema: str
-    rules: str
-    conversation_flow: Optional[str]
-    conversation_context: str
-    resource_constraint: Optional[ResourceConstraint]
-    service_id: str = "gc_main"
