+++ libraries/python/skills/skills/form-filler-skill/form_filler_skill/guided_conversation/__init__.py
@@ -1,0 +1,9 @@
+import logging
+
+from .guided_conversation_skill import GuidedConversationSkill
+
+logger = logging.getLogger(__name__)
+
+__all__ = [
+    "GuidedConversationSkill",
+]
