+++ libraries/python/skills/skills/form-filler-skill/form_filler_skill/guided_conversation/tests/test_integration.py
@@ -1,0 +1,21 @@
+def test_integration():
+    # TODO: This is the next piece of work... getting this all working.
+
+    # language_model = "gpt-3.5-turbo"
+
+    # definition = GCDefinition(
+    #     artifact_schema="schema",
+    #     rules="rules",
+    #     conversation_flow="flow",
+    #     conversation_context="context",
+    #     resource_constraint=None,
+    # )
+
+    # skill = GuidedConversationSkill(language_model=language_model, definition=definition)
+
+    # runContext = RunContext()
+
+    # skill.conversation_init_function(context=runContext)
+    # finished, artifact = skill.conversation_step_function(runContext, "Hello!")
+
+    pass
